unit09/chinook/requirements.txt
streamlit==1.44.1
langchain==0.1.12
openai==1.12.0
python-dotenv==1.0.1 
langchain-openai==0.1.12
langchain-community==0.1.12